Split some functions into a utils module. Updated tests appropriately. Merged test_auxilary_functions into test_utils
